workshop/my-dlt-pipeline/pyproject.toml
[project]
name = "my-dlt-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "altair>=6.0.0",
    "dlt[workspace]>=1.22.1",
    "duckdb>=1.4.4",
    "ibis-framework[duckdb]>=9.0",
    "marimo>=0.20.2",
]
